Move utils to examples
