+++ copulas/__init__.py
@@ -5,3 +5,7 @@
 __author__ = 'MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com',
 __version__ = '0.1.1-dev'
+
+import numpy as np
+
+EPSILON = np.finfo(np.float32).eps
