+++ examples/example.py
@@ -13,5 +13,3 @@ if __name__ == '__main__':
     gc.fit(data)
     LOGGER.debug(gc.sample(num_rows=1))
     LOGGER.debug(gc.cov_matrix)
-    # LOGGER.debug(gc.get_pdf(np.array([1,5,9])))
-    # LOGGER.debug(gc.get_cdf([2, 5, 8]))
